AI projects/Handwritten Notes OCR System/requirements.txt
opencv-python
numpy
reportlab
streamlit
pillow
transformers
torch
